+++ src/main/java/chiamaka/ezeirunne/bookstore/dto/requests/UpdateBookDto.java
@@ -3,14 +3,15 @@ package chiamaka.ezeirunne.bookstore.dto.requests;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
+
 
 @Getter
 @Setter
 public class UpdateBookDto {
     private long id;
     private String title;
-    private BigInteger price;
+    private BigDecimal price;
     private String author;
     private String isbn;
     private int quantityOfBooksAvailable;
